Move createLoanModalId to the loans group modal component
